requirements.txt
streamlit==1.32.0
groq==0.1.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11